fix(yt): replace ytdl-core-discord with ytdl-core
The package seems to have disapeared from GitHub and is currently broken. Replace is using ytdl-core
directly and most of what the module was doing, except we only play opus/webm streams. To compensate
for the change, check the video on add and throw an error if an opus stream doesn't exist.

BREAKING CHANGE: Limited number of video formats that are playable to keep code simple