feat(playlist): add commands to create and play playlists
In some D&D sessions, I've had my music planned out and would like to just play those songs without
having to deal with it too much. This introduces the idea of playlists, which is a list of songs
that will play in order (or you can shuffle them).

BREAKING CHANGE: Migration for adding a playlists table