feat: rewrite in go
The new JS libraries for voice are hard to use and every update seems to require more changes. I
went with Go since I know the language and the library seems to be fairly active/popular and the
voice helper library made it easy to play local files. I also removed a lot of functionality to get
it working. Right now you give it a directory and it will play the files in that directory.

BREAKING CHANGE: runs in different language. no longer includes Docker image